fix: Prevent IndexOutOfBoundsException on text long press (#1561)
There's a bug on Android before O_MR1. Long pressing to select text,
then long pressing again on the textview can crash with this:

```
java.lang.IndexOutOfBoundsException: setSpan (-1 ... -1) starts before 0
  at android.text.SpannableStringInternal.checkRange(SpannableStringInternal.java:434)
  at android.text.SpannableStringInternal.setSpan(SpannableStringInternal.java:155)
  at android.text.SpannableString.setSpan(SpannableString.java:46)
  at android.text.Selection.setSelection(Selection.java:76)
  at android.widget.Editor$SelectionModifierCursorController.resetDragAcceleratorState(Editor.java:5462)
  at android.widget.Editor$SelectionModifierCursorController.onTouchEvent(Editor.java:5316)
  at android.widget.Editor.onTouchEvent(Editor.java:1364)
  at android.widget.TextView.onTouchEvent(TextView.java:8451)
  at app.pachli.core.ui.ClickableSpanTextView.onTouchEvent(ClickableSpanTextView.kt:301)
  at android.view.View.dispatchTouchEvent(View.java:9943)
  ...
```

`resetDragAcceleratorState()` calls `TextView.getSelectionEnd()`, and
even though text is selected at that point `getSelectionEnd()` returns
`-1`.

This `-1` is eventually passed as the `start` value to `checkRange`,
which throws the exception.

Work around this by either resetting the selection if present, or
resetting the text.