feat: Speed up account switching if the user has no account filters (#1564)
Previous code always fetched the user's list of followed accounts as
part of refreshing the local account data when loading the app or
switching accounts.

This may take a lot of time, and is only necessary if the user has the
notification or conversation account filters set to filter accounts the
user is not following.

Change this so the fetch still happens, but only blocks the refresh if
the user has one of those filter options set. Otherwise the fetch
happens in the background.

See https://github.com/pachli/pachli-android/issues/1556