refactor: Don't destructure large classes
Destructuring (`val (_, _, x, y, _, _, z, _, _) = someObject`) is
dangerous, because if the order of the properties in `someObject` change
without changing the types you have a runtime bug instead of a compile
time error.

Replace this with explicit variable creation and assignment for each
used property. Noticed some places where the actionable status is the
correct choice, update those call sites.