feat: Restore the user's reading position in lists (#1548)
Previous code restored the reading position only for cached timelines,
which is currently only the home timeline, and the notifications
timeline.

Extend this to lists. To do this:

- Extend the timeline model to include the "key" that's used to record
the user's reading position. For lists this includes the list's server
ID. For non-restorative timelines this is null. This indicates whether
or not a timeline's position can be restored.

- Move code for saving/reading the status ID at the position to
`TimelineCases` from `CachedTimelineRepository` so it's available in
other timelines.

- Move code for exposing the initial status ID to `TimelineViewModel`
from `CachedTimelineViewModel` so it's available in the other view
models.

- Extend `NetworkTimelineRemoteMediator` to support loading a page
*around* the status at the user's reading position for timelines that
support that.

- Adjust `TimelineFragment` to decide whether to restore the reading
position based on the viewmodel's `initialRefreshKey`, not the type of
the viewmodel.