fix: Delete cached notifications when deleting server notifications (#1585)
Previously, when the user decided to clear all notifications they were
deleted from the server but the local cache was not cleared (because the
remote mediator returned early if they were no results on a refresh), so
Pachli continued to show the (now deleted) notifications.

Fix that -- always delete the notifications.

While I'm here, make the "enabled" state of the "Clear notifications"
menu item contingent on whether there are any notifications available to
delete, and update CachedTimelineRemoteMediator and
ConversationsRemoteMediator to behave the same.

Fixes #1582