refactor: Keep network types out of the rest of the code (#1581)
Previous code defined network types (i.e., the shape of data returned
from the Mastodon API) and embedded it throughout the rest of the code.

This tight coupling is problematic:

- Decisions (types, nullability, data organisation) in the network API
permeate the codebase.
- It pushes code that deals with non-Mastodon servers out of the
network and data layer and into the higher layers.
- The network types had functions unrelated to the network layer.

New features were migrating away from this (e.g., `suggestions`) but
most of the code still had this problem.

Fix this everywhere.

Create new types in `core.model` for the app's data, and use these types
in data/domain/view layers of the code.

In most cases these types are similar to the network types, but breaking
the link **and** evolving the types was too invasive for a single PR.
Some of the properties of the new types that should definitely not be
null have been marked non-null, which allows for instances of `?:`,
`?.`, and `!!` to be removed from calling code.

Move some types from `core.data.model` to `core.model`.

Mapping between data types is done by:

- Network types have `.asModel()` functions, which convert to the
equivalent `core.model` class.
- Other types have `.asEntity()` or `.asNetworkModel()` functions (or
extension functions) to convert to database entities or back to the
`core.network.model` types as necessary, for consistency. Existing
`.from` functions that did this have been removed.