feat: Show emojis grouped by category, and searchable (#1624)
Previous code presented emojis as a flat list, sorted by emoji
shortcode. On servers with many emojis this was difficult to find the
one you wanted, and emojis from different categories were mixed
together.

Fix this.

Display emojis grouped by category, with an additional view for the
category header. Provide a `TextView` to allow the user to search the
emojis by shortcode or category, filtering the list to just the matched
entries. Scroll the list vertically instead of horizontally, with
scrollbars.

Fixes #1226