change: Remove custom @Json custom names from most core.model classes
These were a holdover from when these were the network classes and
serialised directly to the database, and can be removed now.

The exceptions are classes where the database holds user data and is not
just a cache of remote data (e.g., `DraftEntity` references `DraftAttachment`
`NewPoll`, both of which may already be serialised to the database).

For the classes where this can be done perform the removal. Because the
database cache may now contain data in the format, clear the affected
cache tables.