fix: Remove custom @Json custom names from most core.model classes (#1632)
These were a holdover from when these were the network classes and
serialised directly to the database, and can be removed now.

The exceptions are classes where the database holds user data and is not
just a cache of remote data (e.g., `DraftEntity` references
`DraftAttachment` `NewPoll`, both of which may already be serialised to
the database).

For the classes where this can be done perform the removal. Because the
database cache may now contain data in the obsolete format clear the
affected cache tables. This also fixes a potential crash if any of the
stale data is loaded (e.g., reading obsolete data from `StatusEntity`
when loading a thread).