feat: Use the new emoji picker for emoji reactions on announcements
Previous code used the old UI for displaying emoji picker reactions, which
didn't allow searching or categorisation.

Fix this.

Move the emoji UI related code and resources to `app.pachli.core.ui.emoji`,
and create a new `EmojiPickerView`, which wraps the grid display and the
filter text box in to a single view.

Move the configuration properties in to this view, and hide `EmojiAdapter`
as an internal implementation detail. Update `ComposeActivity` and resources
to reflect the new API.

Add `ChooseEmojiDialogFragment` to show a dialog containing `EmojiPickerView`.
Update `AnnouncementsActivity` to show this dialog and use the result when
voting.