fix(deps): update agp to v8.11.0, lint to 31.11.0 (#1641)
The new `MemberExtensionConflict` lint errors appear to be false
positives, tracked at https://issuetracker.google.com/issues/427761232
and should be resolved in the next Android Studio release.

I'm adding them to the baseline (instead of ignoring by ID) to ensure
that future warnings of this type that aren't false positives are still
flagged.