fix: Don't crash when managing accounts on older API devices (#1646)
API 24 devices (at least, possibly some newer API levels) don't want
`android:background` to be a color statelist, crashing because the XML
resource inflater expects to find a drawable instead of a colour.

Work around this by setting the colour to the surface colour, and using
the statelist as the background tint.