feat: Implement edge-to-edge support (#1650)
Edge-to-edge support is enabled for all devices and all activities.
Some activities are fully edge-to-edge (e.g., viewing zoomed in media),
others are partially edge-to-edge (e.g., recyclerview lists running
under the bottom navigation bar), and some are not because the layout
requires fixed buttons at the bottom of the screen.

The main implementation is in `fun View.applyWindowInsets()`. Called
on a view this sets the insets to be applied on each side (not all
sides may have an inset), whether to consume the insets, a mask for
the inset types (systembar, display cutout, etc), and a flag for
whether the insets should track the IME.

This remembers the initial padding and margin of the view, then
defines a listener for new insets that applies them relative to the
original paddings and margine. The insets are either fully or partially
consumed as appropriate.

The listener is set as an apply-window-insets listener and a
window-insets-animation callback.

A number of additional view extensions are defined to provide default
margin or padding inset preferences for different types of views.

Most activities follow the same pattern; specifically, modify
`onCreate()` and:

- Call `enableEdgeToEdge()` before `super.onCreate(...)`
- Call `ViewGroupCompat.installCompatInsetsDispatch(binding.root)`
immediately after `super.onCreate(...)`
- Configure the layout's appbar with default insets
- Configure the layout's toolbar to fade the app bar as it scrolls
- Configure other views (typically `RecyclerView`) with default insets

If a fragment is modified that's typically to apply default insets to
its `RecyclerView`.

Some activities don't fit this pattern because their views are more
complicated or require additional handling.

Set `compileSdk` and `targetSdk` to 36.

Update `TouchImageView` to v3.7.1, it contains a necessary bug fix
for edge-to-edge support.

## Notable user visible UI changes

- In screens that scroll, if there are no UI elements towards the
bottom of the screen, the content now goes under the bottom
navigation bar, with enough padding to clear the bar when the content
is scrolled to the top.

- When scrolling, the app bar, if present, fades as it scrolls out
of view.

- Images and videos are now displayed in immersive mode, with no
system navigation. When zoomed in images can display on the entire
screen, including overlapping any display cutouts for cameras. This
is OK because the user can scroll around to see more of the image.

- Images are loaded at their full size in order to see more detail
when zoomed in.

- When scrolling the screen for an account the account's avatar,
display name, and username fade out in the large header and fade in
on the small header.

- Showing / hiding the keyboard (e.g., when writing a status) should
smoothly slide the UI elements out of the way as the keyboard scrolls
into place, instead of jumping to place while the keyboard scrolls
into an empty spot.