fix: Don't crash unregistering AudioBecomingNoisyReceiver (#1652)
Previously the `AudioBecomingNoisyReceiver` was registered in the
fragment playing the video.

This works for statuses from Mastodon which can only have a single video
attached. However, statuses from other networks may have multiple videos
attached. The receiver would be registered by each video playing
fragment and then unregistered when they pause/stop. Because the
receiver is a singleton the second attempt to unregister it (if the
status has two or more video attachments) would crash.

Fix this by moving the broadcast receiver to `ViewMediaActivity`.
Implement the receiver as lifecycle observer so it can
register/unregister itself without needing to modify `ViewMediaActivity`
beyond instantiating the object.

Update `ViewMediaAdapter` to include an `onAudioBecomingNoisy` method to
call on receipt of the broadcast.

Update `ImagePagerAdapter` to implement `onAudioBecomingNoisy` and call
an identically named method on each fragment.

Implement `onAudioBecomingNoisy` in `ViewVideoFragment` and pause the
player.