feat: Pause video playback when swiping through media (#1653)
Previous code didn't pause video playback when swiping through media.
This isn't a problem on Mastodon where a status can't mix and match
video and image attachments, but it is a problem from posts with other
networks that can. If video was playing it would continue playing after
the user swiped to a different attachment.

Fix this. Provide a new method, `onDetachedFromWindow`, that the media
fragments can implement, and call this method from `ImagePagerAdapter`
when `ViewPager2` detaches the fragment after a swipe. Pause the player
when this happens.