fix: Don't crash with network activity on the main thread sharing media (#1660)
Previous code made the network request on `Dispatchers.IO`, but then
read the response body on `Dispatchers.Main`. If the response was large
and the body hadn't been fully read in the request this could cause a
crash with a `NetworkOnMainThreadException`.

Fix this by splitting the "share" operation in to `shareMedia` which
runs on the main thread and manages the UI, and `downloadUrlToTempFile`
which performs all the disk and network operations on `Dispatchers.IO`.