fix: Prevent NPE where *TimelineViewModel.repository might be null (#1671)
Previous code could crash in e.g,. `NetworkTimelineViewModel` at a call
site like this:

```kotlin
override fun handleFavEvent(favEvent: FavoriteEvent) {
    viewModelScope.launch {
        repository.updateActionableStatusById(favEvent.statusId) {
            it.copy(favourited = favEvent.favourite)
        }
    }
}
```

where `repository` was null.

This happened because `repository` was a property of
`NetworkTimelineViewModel` but `handleFavEvent` (or other functions)
could be called by a flow collector running in `TimelineViewModel.init`.

So `NetworkTimelineViewModel.handleFavEvent` might be called before
`NetworkTimelineViewModel` was fully constructed, and `repository` was
null at that point.

Fix this by exposing `TimelineViewModel.repository` to subclasses. To
make this work the subclass has to pass the concrete type of the
repository as a type parameter to `TimelineViewModel`