fix: Reduce network calls when filtering notifications
Previously, all notifications were fetched from the server and filtering
was performed client side.

If the user selected a type of notification that was not present in a
page then then the next page would be fetched. If it was not present in
that page the next page would be fetched, and so on, until enough
notifications of that type had been fetched from the server.

If the notification type was rare (e.g., a moderation action) this could
end up paging through all the user's notifications with very few results,
burning through API quote and causing the notification fetch to take a
very long time.

The UI was also very jumpy while this happened.

Fix this, pass the set of notification types to filter down to
`NotificationsRemoteMediator` and pass that as an API parameter. The
results are returned much faster, and without UI jumping.