feat: Show role badges in timelines and account lists
Previously any roles for the account were only shown on the account page,
and server-side colour of the role was blended with the account page.

Now, show the role badges above the account info in the timeline and when
showing lists of accounts (e.g., autocomplete, searching, or managing lists).

The badge is shown apart from the user generated content so it can't be
spoofed with e.g., an emoji.

The colour blending could give some terrible results, so remove that
entirely.