fix: Prevent crash when replying from notifications (#1692)
The change that migrated to `core.model` classes everywhere missed the
intent created when tapping a "quick reply" on a notification. The
status' visibility was going in to the intent as a
`core.model.Status.Visibility` and being retrieved as a
`core.network.model.Status.Visibility`, causing a class cast exception.