fix: Don't crash on older devices loading role chip resources
try-with-resources and `use` doesn't work on a `TypedArray`, the
desugaring works on `AutoClose`, not the `AutoClosable` that Kotlin
uses.

This crashes with:

```
java.lang.ClassCastException: android.content.res.TypedArray cannot be cast to java.lang.AutoCloseable
```

The fix is to import `androidx.core.content.res.use`, which overrides
`use` from the Kotlin stdlib and does the right thing.