fix: Show large/zoomed images with padding to avoid insets (#1695)
Previously you could zoom in on an image and it would appear
edge-to-edge in immersive mode and you can scroll around. But the edge
of the image might still appear under the camera cutouts in the display.

Add padding; the image can still be scrolled under the camera cutout,
but the edge of the image can be moved away from it completely.