build: Add a lint check for TypedArray.use (#1702)
`TypedArray` implements `AutoCloseable`, but doesn't have a desugared
`close` implementation on older devices. This causes a runtime exception
on older Android devices.

Typically this is triggered by using the `kotlin.use` extension method
on something that returns a `TypedArray`.

The fix is to call `androidx.core.content.res.use` instead, which can
normally be done simply by adding the import for it.

Add a lint check to detect this and suggest the change.

While I'm here, simplify the resource references in ProfileChip.kt.