feat: Restore reading position on Local and Federated timelines
Previous code reloaded the Local and Federated timelines from the latest
post when the app starts and/or the viewmodel that cached the relevant
flow was discarded.

Fix this; remember the reading position for both timelines and restore
it when loading the timeline.

High level changes include:

- `NetworkTimelineRepository` now emits a
`PagingData<TimelineStatusWithAccount>` to ensure up to date view data
is always included.

- The `PageCache` now uses a `Mutex` to guard access instead of using a
`synchronized` block; this ensures it works with coroutines as well as
threads.

- Position restoration happens in fragments when CREATED, so it happens
once (previous was STARTED, approximately every time the Fragment resumed).

- `NetworkTimelineRemoteMediator` is re-written to be simpler, and lock
the `PageCache` for the duration of a load.

- Additional logging.

A number of bugs have been fixed as part of this:

- The "Load Newest" functionality could fail because of the incorrect
application of a null test in `NetworkTimelineRemoteMediator`.

- Muting/blocking accounts were applied on the server, but server
behaviour is asynchronous, resulting in the muted/blocked account still
appearing when the timeline was invalidated. Work around this by caching
data about the mute/block and applying it when the timeline is refreshed.

- Changes to `StatusViewData` (e.g., hiding/showing media) weren't always
reflected in the timeline when it was refreshed.

- The index returned from `getRefreshKey()` for non-cached timelines could
sometimes be wildly off, jumping the user around the timeline.