refactor: Implement StatusRepository in other repositories (#1708)
Previous code that operated on timelines and statuses generally injected
a repository for the timeline and `StatusRepository`, then had to
remember which repository to use for which operation. In some cases the
timeline repository included methods that just wrapped those in
`StatusRepository`.

Simplify this. Now the goal is that viewmodels work with a single
repository for the timeline they operate on. The `StatusRepository`
methods have been moved to an interface, implemented by
`OfflineFirstStatusRepository`.

The timeline repositories inject `OfflineFirstStatusRepository` as a
constructor parameter and implement `StatusRepository` by delegating to
the injected `OfflineFirstStatusRepository`.

Most viewmodels have been updated to do this. Some legacy ones which
don't have a specific timeline repository are left unchanged. When those
are converted `StatusRepository` can be marked internal to `core.data`.