fix: Show animated avatars at correct size in timelines
Previous code applied a background to avatars to ensure that partially
transparent areas are fully opaque when drawing content under them. This
happened before the avatar image was scaled for use.

Not sure why, but this has started to cause animated avatars to be
displayed at the wrong size.

Fix the problem by only performing the compositing after the avatars have
been scaled.

While I'm here, clean up an unnecessary pass of "null" as the value.