fix: Always load cached thread status data with the correct account
When viewing a thread the previous code loaded the status, including the
view data, from the local database, but without using the Pachli account
ID.

This could result in the view showing status view data from another
account.

For example, consider two accounts on the device that both have the same
status in their timeline. The status is in a different language, and the
first account translates the status. This is now part of the cached view
data.

The second account sees the same status, does not translate it, and clicks
through to view the thread. The cached data is loaded, and that's the
translated status -- so the second user sees the translation even though
they did not explicitly request it.

Fix this by passing the Pachli account ID through to the cache query and
using it in the query.