feat: Allow delete over autocomplete-inserted space
Previous code inserted included a space at the end of the text when as
part of the autocomplete result when insertinga suggestion (account,
hashtag, or emoji).

This meant the space was considered part of the autocomplete result. So
if the user autocompleted e.g., a hashtag, and wanted to delete the
inserted space to replace it with a comma, the autocomplete portion of
the result would also be deleted.

Change this behaviour, and insert the space after the autocomplete result
is inserted, separate from the result itself. This allows the user to
delete the space without affecting the autocomplete result. Pressing
delete a second time will delete the autocomplete portion of the result.