refactor: Include the colons in an emoji's short code
Previous code used the emoji data exactly as received from the server.
This replicates a Mastodon API bug, where the network model for an emoji
does not include the `:` in the emoji's shortcode.

Fix this. Include the colons when converting from the network model to
the app model, and update code that inserts emojis so it doesn't need to
include the colons.