feat: Allow the user to see an emoji's detail
Previous code shows the emoji grid and allowed the user to long-press an
emoji to see the shortcode. But the emoji were still quite small.

To improve this, add a "Show detail" checkbox next to the emoji filter.
When checked a "detail" panel appears on the right and the emoji tap mode
changes.

With the panel closed tapping an emoji inserts it in the text. With the
panel open tapping an emoji shows it at roughly 4x size in the detail
panel, and the emoji shortcode is shown below the image. The user can tap
the detail image to insert the emoji.