feat: Show emojis as images when writing/editing posts (#1747)
Previous code showed the user the emoji shortcode when inserting an
emoji, whether through autocomplete or the emoji picker.

Change that -- insert the shortcode, but then wrap the shortcode in an
`EmojiSpan` to display the actual emoji image.

When editing posts run the text-to-edit through `emojify` first so it
also has the emoji span's inserted.

There's a bug in `EditText` that causes the emoji images to not always
appear when added. To work around that set
`android:layerType="software"` (see
https://github.com/koral--/android-gif-drawable/issues/368).