wip: Experiment with Jetpack Compose
Try and convert the translation model manager UI to Compose to get a feel
for the complexity and what features are currently missing.

Android Studio does not like this, see this bug for details:
https://issuetracker.google.com/issues/441253579

Stuff still to do to take this forward:

- No scrollbar on the list
- Haven't finished checking all the edge-to-edge stuff
- Haven't converted the dialogs

Also, all the build.gradle.kts dependency stuff is janky and copied from
the "Setting up Compose" docs, rather than integrating it nicely into
the build system.