fix: Periodically update timestamps in threads, search results (#1769)
If "Use absolute time" is off then the status date/time is shown
relative to the current time, and updated every minute. The update is
performed by notifying the adapter the item range has changed with a
dedicated payload to indicate the update time should be re-bound.

This wasn't enabled for threads or search results. Implement the code to
do that.

While I'm here clean up the payload handling. The payload can be be any
arbitrary object. Make sure it's consistently treated as a
`List<List<Any?>>?`.

This is difficult to make properly safe, as `RecyclerView` just assumes
it's `Object` and doesn't parameterise this.

Pachli uses `List<List<Any?>>?` because the call to
`DiffUtil.ItemCallback.getChangePayload()` may return a list of items,
if multiple parts of the status or other item have changed (e.g., if the
text of the status changed but the attachment did not there is no need
to re-fetch and decode any image attachments on the status). So
`getChangePayload()` has to return a `List<Any?>?`, and this return
value is then put in another `List` when passed as the parameter to
`onBindViewHolder`.