refactor: Prepare for blur filter support
Get ready for blur filter support and clean up some of the content filter
UX handling.

- Filters are included in the account data, so remove the obsolete code
for dealing with the possibility that the filter could not be loaded. This
simplifies the code that manipulates the filter.
- Provide a flow that contains the server-supported filter actions. Use
this in the UI to control the visibility of the different radio buttons.
- Implement `isDirty` and `validationErrors` as flows that map over
`contentFilterViewData` instead of calling `onChange` whenever the view
data changes.