feat: Support filters that hide media attachments (#1776)
Mastodon 4.4.0 introduced a new filter action, `blur`. Misnamed, if
triggered this should show the text of the status but not show the
attachments. The user should be able to click the obscured attachments
to view them. The filter applies to all attachments on the status.

Implement support for this.

When editing filters, if the server supports the `blur` filter action
then show an extra radio button for the user to choose this option.

The changes to support displaying statuses that match this filter are
more complex.

Previously, displaying attachments depended on on whether the status
was marked sensitive, and if it was, whether the user had chosen to
view sensitive content by default.

An attachment may now be hidden for multiple reasons. Model this with
the new `AttachmentDisplayAction` subclasses (`.Show` and `.Hide`).
The `.Hide` subclass contains a new `AttachmentDisplayReason` with
details about why the attachment was hidden.

The action is stored in `StatusViewData.attachmentDisplayAction`,
replacing the simple boolean `StatusViewData.isShowContent`.

Make StatusViewData binding more efficient with a new change payload
that indicates the attachments, or the attachment display action,
have changed. If this is not present then the attachments are not
reloaded/redisplayed.

To keep this consistent move the `DiffUtil.ItemCallback<StatusViewData>`
to a static object that can be reused anywhere `StatusViewData` is
displayed in a list.