refactor: Don't start activities from StatusBaseViewHolder
Previous code started activities from `StatusBaseViewHolder` if to show
preview card authors and view media in preview cards.

This is a layering violation. Amend `StatusActionListener` to account for
these, and call the listener (which will eventually be the fragment or
activity hosting the `StatusBaseViewHolder`). Let that fragment or activity
start the next activity.