refactor: Implement ReportViewModel as a subclass of NetworkTimelineViewModel (#1816)
This is prep work for a larger change to have the report UI follow the
new filter decision API. Building `ReportViewModel` on
`NetworkTimelineViewModel` reduces the amount of duplicated code.

`ReportViewModel` uses `@AssistedInject` and the other
`TimelineViewModel` subclasses don't, so change them (and the code that
creates them) to use that instead of the previous approach of storing
the timeline in the `SavedStateHandle`.