refactor: Show attachment previews in new AttachmentPreviewView (#1819)
The attachment preview is the view that shows:

- The attachment's preview image (or placeholder, depending on user
preferences).
- The playback overlay indicator, if the attachment is playable.
- The "ALT" badge at the bottom right if the attachment has a
description.

Previously, the logic to display this was scattered across multiple
places, including `StatusBaseViewHolder`, `StatusViewHolder`, and
`StatusViewHelper`.

Fix this.

Create a proper Android compound view, `AttachmentPreviewView`, which
encapsulates the `MediaPreviewImageView`, the `TextView` for the badge,
and the logic for loading the attachment preview (or flat colour /
blurhash), displaying the overlay, and showing the badge.

Use this view in layouts, and adjust the code that uses those layouts to
call the methods in `AttachmentPreviewView` to bind the `Attachment`
data to the view.