feat: Improve accessibility of account name, handle, and roles
Provide a consistent `contentDescription` for account name, handle,
and roles, in `TimelineAccount.contentDescription`.

This ensures any dots in the name are read correctly (e.g., the name
"Mastodon.social Staff" is now read as "Mastodon dot Social Staff", not
"Mastodon <pause> Social Staff". Some TLDs TalkBack doesn't read
correctly, like ".xyz", are special-cased.

Mark the chip group that contains any account roles as unimportant for
accessibility, as the information is included in the account's
`contentDescription`. `ChipGroup` overrides this, so provide a custom
`RoleChipGroup` that does the right thing.