refactor: Move attachment display functionality to new `AttachmentsView`
Previous code duplicated functionality to display attachments in
`StatusBaseViewHolder` and `StatusViewHelper`. As well as the code to
display the attachments this also duplicated the code to decide whether
to hide them, and the UI controls for showing and hiding attachments.

There was also this weird thing were the attachment grid exposed the
individual image views and it was the responsibility of the higher level
code to load the images.

Scrap that, and re-implement as a hierarchy of custom views.

The bottom level layout is `AttachmentPreviewView` (created in a previous
change). This displays the attachment preview, the "playable" overlay,
and the "ALT" badge (if appropriate).

These are organised into a grid with the new `AttachmentGridView`. This
lays out the `AttachmentPreviewView` into a grid and binds the click
listeners. This both simplifies and extends the previous implementation,
`MediaPreviewLayout`.

This is coordinated with a new `AttachmentsView`. This holds the
`AttachmentGridView` as well as the views to show/hide the preview UI
with appropriate warning messages, and the views to use if the user
chose not to load images.