refactor: Extract the post controls to a dedicated compound view
The logic and UI for the strip of controls below a post (reply, reblog,
etc) was tightly coupled with `StatusBaseViewHolder` and its subclasses.
This makes the code longer and more complicated, increases duplication
in code and layouts, and increases the amount of work to do to display
statuses in other contexts (e.g., as part of a quote).

Extract this into a new `StatusControlView` compound control that can
be used only in places where acting on the status directly is necessary.