feat: Allow user to drag attachments to reorder when composing (#1846)
Show a drag-handle below each attachment preview if there is more than
one attachment loaded.

Dragging the handle, or long-pressing on the attachment item starts a
drag. Dropping the attachment in a different spot in the list reorders
the list.

Add "Move up" and "Move down" entries to the per-attachment menu so
there is an accessible way to access this functionality

Fixes #1187