fix: Don't show "Delete conversation" menu item outside ConversationsFragment
Previous code showed the menu item for any statuses with `direct` visiblity.
Since these messages can appear in other timelines (the home timeline,
notifications, in threads, etc) the menu item would appear but the
hosting fragment did not support deleting the conversation.

This triggered an assertion and a crash.

Fix this by only showing the menu if the viewdata is `ConversationViewData`,
which limits it to `ConversationsFragment`.