fix: Update counts when acting on a status
Previously, when the user favourited or reblogged a status the associated
property was updated, but the associated count wasn't. So the timeline could
show the user had favourited a status, but also the status had 0 favourites.

In addition, the opportunistic update was wrapped in a database transaction
that only completed when the network operation completed, so it was slower
than necessary, *and* the result of the network operation wasn't used to
update the local cache.

Fix this.

1. The update operations are no longer wrapped in a transaction, so the
opportunistic updates are reflected more quickly in the UI.

2. When favouriting or reblogging the associated count is also incremented
(or decremented, if appropriate) as part of the opportunistic update, so
the visible counts will also change.

3. The `Status` returned by the operations is used to update the local
cache, so when the operation completes the `Status` data displayed to the
user is completely refreshed, not just the data of the operation they
performed.