fix: Exclude boosted posts when reporting an account (#1858)
When reporting an account the user is prompted to select one or more
statuses sent by the account being reported.

This was including statuses reblogged by the account, without any
indication they were reblogs.

Fix this by explicitly excluding reblogs from the timeline when fetching
statuses to include in a report.