fix: Ensure "Open in browser" is safe when viewing a thread (#1860)
Previous code had two problems.

1. A mis-match between the ID of the status used to root the thread and
the URL used if "Open in browser" was used. Worst case one of them could
have been the status and the other could have been a reblog of the
status, so there's no risk the user would see different content.

Fix this by always using the ID and URL of the actionable status.

2. It's technically possible for a Status from the API to not have an
associated URL. `ViewThreadFragment` didn't consider this, and would
crash if (a) a null URL was provided, and (b) the user clicked the menu
item.

Fix this by gating display of the menu on whether or not the URL is
null.