fix: Prevent toolbar flickering when scrolling (#1861)
The `liftOnScroll` mechanism creates flicking effects, especially in the
light UI, and doesn't signal anything that's not also signaled by the
scrollbar position, missing toolbar header (if scrolled), and/or content
clipped off the top of the screen.

Remove it. However, even when removed it can still cause some
flickering, so set an explicit colour with a style.