fix: Show reported post when reporting an account
This used to be the behaviour, and was dropped in the refactoring in
49ccf30ca32965c3e0202f762125037ac9ac8f9d.

To fix this:

- Make `Timeline.User.Replies` restoreable.
- Set the desired reading position in `ReportActivity`

This surfaced a bug in the position restoration code for network
network timelines. Previously, if the `state` passed to `getRefreshKey`
was null the ID in middle of the first loaded page was used.

This might not work -- testing showed the desired key might be in the
*second* page loaded. The `state` passed to `getRefreshKey` doesn't have
the initial key that would be passed to `load`, so pass a copy of the
initial key in the constructor (it may still be null), and fall back
to that if the `state` is null.