fix: Don't show unavailable accessibility actions
Previous code offered the user accessibility actions on a status that
might not be valid.

For example, in a notification about someone boosting one of your posts
the user was offered the chance to boost the boost. That doesn't work.

To fix this, since `StatusControlView` already knows which actions are
allowed, amend it to expose that as a set of actions.

`ListStatusAccessibilityDelegate` pulls this from the `StatusControlView`
(if present), and uses that to conditionally enable some actions;

- reply
- reblog
- favourite
- bookmark
- more