refactor: Extract code to display statuses to reusable components (#1864)
Previous code had most of the status display code in
`StatusBaseViewHolder`, with specialised viewholders per type (detailed,
conversation, notification, etc).

This was a problem for several reasons:

- The different view layouts for each viewholder duplicated a lot of the
common layout for a status. Keeping them in sync was difficult, and
caused confusion.
- The view holder had to handle the "top-line" info for the status, the
status itself, and the status controls, increasing complexity.
- The "Report account" code used a completely different bit of code, in
`StatusViewHelper`. This also had to be kept in sync with any changes.

All of this is a problem if a status is going to be shown in any
additional contexts -- for example, as a nested quote inside another
status.

So fix it.

Responsibility for displaying the body of a status (avatar, content,
polls, attachments, preview cards) in different contexts is now in the
following compound controls.

- (abstract) `StatusView`: Base class, provides the core behaviour, does
not know the specific layout. Each subclass has to assign views to
properties for `StatusView` to operate on.

- TimelineStatusView: Displays the body of a status in a timeline.

- DetailedStatusView: Displays the "detailed", or focused status in a
thread.

- ConversationStatusView: Displays the most recent status in a
conversation.

- NotificationStatusView: Displays a status embedded in a notification.

- ReportStatusView: Displays a status as part of reporting an account.

The layouts for the different viewholders are reimplemented to contain
these new compound views, the top-line view (if any), and the status
controls (if any). This removes most of the logic, and complexity from
the viewholders.

This does cause a problem with aligning `StatusControlView` in the
layouts. Previously this could be done by aligning against views that
were part of the status body, but that's not possible with compound
views.

To fix this, use the existing dimension resources that specify things
like avatar padding/margin, and width, and use databinding to support
calculation syntax in the relevant layouts, like:

```
android:layout_marginStart="@{@dimen/account_avatar_margin + @dimen/timeline_status_avatar_width}"
```

This requires `kapt`, and is a stop gap until all of this is eventually
converted to Compose.