refactor: Prevent nullability issues in NotificationViewData types
Previous code had a single `NotificationViewData` type to cover all
possible types of notifications.

This meant it had to have a number of nullable properties, and a
`type` property. Every time code interacted with `NotificationViewData`
it had to inspect the `type` property and assume/hope the nullable
properties were correctly non-null for that particular type of
notification.

Improve this by converting `NotificationViewData` to a sealed interface.
Each distinct notification type is a class that implements this interface
and just the (non-nullable) properties it needs.

Add a `NotificationViewData.WithStatus` interface to cover the multiple
notification types that have an embedded `StatusViewData`. This implements
`IStatusViewData` by delegating to the embedded `statusViewData`, and
allows e.g,. the accessibility to exclusively operate on types that
implement `IStatusViewData`, removing additional code complexity in that
class.